[BU000003ZB1WH5] Improve handling of boolean query parameters on GET requests (included in v1.1.0)
PHP is super weird in how it casts booleans to strings:

```php
print(var_dump(strval(true)));
// string(1) "1"
print(var_dump(strval(false)));
// string(0) ""
````

This causes problems in the library when you try to filter a list resource with a boolean attribute, since we end up adding either `&enabled=1` or `&enabled=` to the URL, when the API expects strings of "true" or "false". Other libraries handle this okay, but PHP is the odd one out.

This allows users of the PHP library to continue passing booleans, which makes sense, but casts them correctly before the request is made.