Standardise how we expect exceptions in our tests
PHPUnit offers two ways to expect an exception in your tests:

* Annotations on the test method, e.g.:

```php
/**
* @expectedException        {{ .Config.module_name }}\Core\Exception\InvalidStateException
* @expectedExceptionMessage Mandate is already active or being submitted
*/
```

* Calling `$this->setExpectedException` in the example

At the moment, we're using both. Let's standardise on the latter
for the sake of convention. I'm choosing that way since it'll
be more familiar to users of other (non-Java!) languages, and it's
the approach mentioned first in the [PHPUnit docs][1].

[1]: https://phpunit.de/manual/current/en/writing-tests-for-phpunit.html#writing-tests-for-phpunit.exceptions