Throw a GoCardlessInternalException when the API returns a 500
`ApiException::getErrorForType($error_type)` is responsible for
working out from the `type` returned in the API error response
what exception should be thrown. In the docs, we [detail][1]
four available error types:

* `invalid_api_usage`
* `invalid_state`
* `validation_failed`
* `gocardless`

At the moment, we support the first three but not the last, in
the final case throwing a `GoCardlessProException` with message
`Invalid error type gocardless`, as [reported][2] by @jmper.

This fixes the issue, instead throwing a
`GoCardlessInternalException` as we ought to, and improves the
tests around this code at the same time.

Fixes #41.

[1]: https://developer.gocardless.com/api-reference/#errors-error-types
[2]: https://github.com/gocardless/gocardless-pro-php/issues/41